requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-login==0.6.2
flask-migrate==4.0.5
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.0.0
google-generativeai==0.3.1
python-jose==3.3.0 